Module Avro.Obj_container_file

The object container file storage format

This format is used to store a collection of rows which have the same type, using an efficient binary encoding.

module Decode : sig ... end

Reading from an input

module Encode : sig ... end

val register_decompression_codec : string -> decompress:(string -> string) -> unit

Register decompression codecs. Defaults are "null" and "deflate".