Module type Output.CUSTOM

Custom output

val small_buf8 : bytes

Small temporary buffer, used for lengths and the likes

val write_byte : char -> unit

Write a single byte

val write_slice : bytes -> int -> int -> unit

write_slice b i len writes the slice of length len of b starting at i.

val flush : unit -> unit

Non specified hint that the data may be flushed onto the disk or network. Doing nothing is acceptable.